Refactored IUpdatableMod.
